test(runs): home the terminal-status guards with the status contract tests
Move both new guards out of test_safe_deploy.py (a deploy-script and
worker-lifecycle file that merely imports OPEN_RUN_STATUS_VALUES for its
pre-swap check) into test_agent_run_state_machine.py, where the contract
tests for brain/systems/runs/status.py already live.

A guard nobody finds when they add a RunStatus member is not a guard.
